# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "Laravel Api.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/signup": {
            "post": {
                "summary": "Register user",
                "operationId": "signup",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "summary": "Login",
                "operationId": "login",
                "responses": {
                    "202": {
                        "description": "accepted"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "summary": "Logout - revoke token",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/auth/add_user": {
            "post": {
                "summary": "Register new user with all needed data",
                "operationId": "addUser",
                "responses": {
                    "201": {
                        "description": "created"
                    },
                    "400": {
                        "description": "validation error"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/auth/fetch_user": {
            "post": {
                "summary": "Fetch users",
                "operationId": "fetchUser",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    }
}